Health bar and combat system
-> Added health bar

-> Added combat system

Description of combat system:
The enemies are idle, moving in random directions after every x seconds, but without leaving their radius ( relatively to their spawn point ). When the player enters the enemy's radius, it starts chasing the player and doubles their combat radius while they are in combat mode. When the enemy is close to the player, it attacks them, deal damage and knockback them. When any character is hit, there is a short animation of the character toggling red/default color of the sprite for a short time. Upon the character's death, it is respawned in the spawn point set.